+++ datanode/sqlstore/sqlstore.go
@@ -88,6 +88,7 @@ var defaultRetentionPolicies = map[RetentionPeriod][]RetentionPolicy{
 		{HypertableOrCaggName: "referral_set_stats", DataRetentionPeriod: "1 year"},
 		{HypertableOrCaggName: "referral_fee_stats", DataRetentionPeriod: "1 year"},
 		{HypertableOrCaggName: "funding_payment", DataRetentionPeriod: "1 year"},
+		{HypertableOrCaggName: "volume_discount_programs", DataRetentionPeriod: "1 year"},
 	},
 	RetentionPeriodArchive: {
 		{HypertableOrCaggName: "*", DataRetentionPeriod: string(RetentionPeriodArchive)},